v1.3.2 make fps drop speed independent of fps

+++ src/main/java/com/xpdrops/XpDropInFlight.java
@@ -13,6 +13,6 @@ public class XpDropInFlight
 	float yOffset;
 	float xOffset;
 	float alpha;
-	int frame;
+	float frame;
 	int hit;
 }
